+++ src/Bounded.hs
@@ -1,13 +1,15 @@
 module Bounded
   ( -- * Bounded
     Bounded(..)
+    -- * Lower
   , Lower(..)
+    -- * Upper
   , Upper(..)
-    -- ** Newtypes
+    -- * Newtypes
   , Bound(..)
   ) where
 
 import Data.Semilattice.Bound
 import Data.Semilattice.Lower
 import Data.Semilattice.Upper
-import GHC.Enum
+import GHC.Enum (Bounded(maxBound, minBound))